% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_manifold.R
\name{compute_manifold}
\alias{compute_manifold}
\title{Compute the manifold of gene expressions}
\usage{
compute_manifold(
  counts,
  method = "SAVER",
  preprocess = TRUE,
  ncores = 1,
  nfeatures = 3000,
  dims = 20,
  neighbors = 20
)
}
\arguments{
\item{counts}{An expression count matrix. The rows correspond to genes and
the columns correspond to cells. It can be sparse.}

\item{method}{Method for estimating the manifold. Can be either "SAVER",
which uses SAVER to recover the true expression, or 'neighbor',
which uses nearest neighbors to approximate the true expression}

\item{preprocess}{Whether the expression count matrix should be preprocessed,
where all genes with zero counts will removed. Default is TRUE.}

\item{ncores}{Number of cores to use. Default is 1. (SAVER only)}

\item{nfeatures}{Number of features to use for FindVariableFeatures. Default is 3000. (Neighbor-based only)}

\item{dims}{Number of dimensions for nearest neighbors. Default is 20.(Neighbor-based only)}

\item{neighbors}{Number of neighbors to output. Default is 20. (Neighbor-based only)}
}
\value{
A matrix of estimated gene expressions.
}
\description{
Compute the manifold of gene expressions
}
\examples{
load('./data/TMS_marrow.RData')
compute_manifold(dataset.counts, method='SAVER')
compute_manifold(dataset.counts, method='neighbor')
}
