% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dispersion.R
\name{estimate_dispersion}
\alias{estimate_dispersion}
\title{Estimate dispersion parameters}
\usage{
estimate_dispersion(
  counts,
  manifold,
  cell.types,
  ages,
  model = "cCV",
  preprocess = TRUE,
  ncores = 1,
  size.factor = NULL,
  cell.types.to.use = NULL,
  ages.to.use = NULL,
  cell.types.cutoff = 10
)
}
\arguments{
\item{counts}{An expression count matrix. The rows correspond to genes and
the columns correspond to cells. It can be sparse.}

\item{manifold}{The computed manifold from compute_manifold. It should be a matrix
of the same dimension as counts.}

\item{cell.types}{A vector with the same length as the columns of counts. Each value should be
a string corresponding to the cell type of each cell in counts.}

\item{ages}{A vector with the same length as the columns of counts. Each value should be
a string corresponding to the age group of each cell in counts.}

\item{model}{Model to be used for estimating dispersion parameters. Can be either "cCV" (constant CV),
"cFF" (constant Fano factor), or "cVar" (constant variance).}

\item{preprocess}{Whether the expression count matrix should be preprocessed,
where all genes with zero counts will removed. Default is TRUE.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{size.factor}{A vector of cell size normalization factors.
Default uses mean library size normalization.}

\item{cell.types.to.use}{A vector of unique cell types to be used in the estimation process.
Default is all cell types.}

\item{ages.to.use}{A vector of unique age groups to be used in the estimation process.
Default is all age groups.}

\item{cell.types.cutoff}{Minimum number of cells in a cell type x age group. Default is 10.}
}
\value{
A named list for all dispersion parameters. Each element is a table
of dispersion parameters for a cell type. The columns in each table contain
dispersion parameters for each age group.
}
\description{
Estimate dispersion parameters
}
\examples{
load('./data/TMS_marrow.RData')
load('./data/TMS_marrow_manifold.RData')

ages <- data.frame(index=dataset.age)
age_map <- data.frame(index=c(1:length(dataset.age.levels)),
age = dataset.age.levels)
celltypes <- data.frame(index=dataset.celltype)
celltype_map <- data.frame(index=c(1:length(dataset.celltype.levels)),
celltype = dataset.celltype.levels)
ages <- join(ages, age_map)
celltypes <- join(celltypes, celltype_map)

estimate_dispersion(dataset.counts, dataset.saver$estimate, celltypes$celltype, ages$age,
                    model = 'cCV', ncores = 4, cell.types.cutoff = 10)

}
