% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tms_marrow}
\alias{tms_marrow}
\title{Tabula Muris Senis (TMS) single-cell RNA-seq dataset}
\format{
An object of class \code{list} of length 1.
}
\usage{
tms_marrow
}
\description{
20,138 genes and 40,220 cells from mouse bone marrow.
}
\references{
Tabula Muris Consortium (2020). A single-cell transcriptomic
atlas characterizes ageing tissues in the mouse. \emph{Nature, 583}(7817),
590–595.
}
\keyword{datasets}
