% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_manifold.R
\name{compute_manifold}
\alias{compute_manifold}
\title{Compute the manifold of gene expressions}
\usage{
compute_manifold(
  counts,
  method = "SAVER",
  preprocess = TRUE,
  ncores = 1,
  nfeatures = 3000,
  dims = 20,
  neighbors = 20
)
}
\arguments{
\item{counts}{An expression count matrix. The rows correspond to genes and
the columns correspond to cells. It can be sparse.}

\item{method}{Method for estimating the manifold. Can be either "SAVER",
which uses SAVER to recover the true expression, or 'neighbor',
which uses nearest neighbors to approximate the true expression}

\item{preprocess}{Whether the expression count matrix should be preprocessed,
where all genes with zero counts will removed. Default is TRUE.}

\item{ncores}{Number of cores to use. Default is 1. (SAVER only)}

\item{nfeatures}{Number of features to use for FindVariableFeatures. Default is 3000. (Neighbor-based only)}

\item{dims}{Number of dimensions for nearest neighbors. Default is 20.(Neighbor-based only)}

\item{neighbors}{Number of neighbors to output. Default is 20. (Neighbor-based only)}
}
\value{
A matrix of estimated gene expressions.
}
\description{
Compute the manifold of gene expressions
}
\examples{
# Loading test data from Tabula Muris Senis
devtools::install_github("fmicompbio/TabulaMurisSenisData")
library(TabulaMurisSenisData)
tms_marrow <- TabulaMurisSenisDroplet(tissues = "Marrow")
tms_marrow_counts <- tms_marrow$Marrow@assays@data$counts
rownames(tms_marrow_counts) <- rownames(tms_marrow$Marrow)
colnames(tms_marrow_counts) <- colnames(tms_marrow$Marrow)
tms_marrow_counts <- as(tms_marrow_counts, "dgCMatrix")

tms_marrow_ages <- which(tms_marrow$Marrow$age \%in\% c("3m", "30m"))
tms_marrow_cell_types <- which(tms_marrow$Marrow$cell_ontology_class \%in\% c('erythroid progenitor',
'granulocytopoietic cell', 'hematopoietic precursor cell',
'megakaryocyte-erythroid progenitor cell', 'precursor B cell'))
tms_marrow_counts <- tms_marrow_counts[, intersect(tms_marrow_ages, tms_marrow_cell_types)]

# tms_marrow_manifold_saver <- compute_manifold(tms_marrow_counts, method = "SAVER")
tms_marrow_manifold_neighbor <- compute_manifold(tms_marrow_counts, method = "neighbor")
}
